Add more OOP to the chat bot program
